// Source/ng_xbrl_sample_app/src/app/model/instance-process.ts
export class InstanceProcess {

  formId: number;
  formName: string;
  taxonomyFile: string;
  dataPeriod: Date;
  version: string;
  processId: number;
  instanceFile: string;
  instanceStatus: string;
  instanceStatusStyle: string;
  instanceCreateBy: string;
  instanceCreateDate: Date;
  validationStatus: string;
  validationCreateBy: string;
  validationCreateDate: Date;
  validationStatusStyle: string;

}